+++ KinoLenta/Common/Extensions/DataExtension.swift
@@ -1,11 +1,6 @@
 import Foundation
-import UIKit
 
 extension Data {
-    func toImage() -> UIImage? {
-        UIImage(data: self)
-    }
-
     var hexDescription: String {
         return reduce("") { $0 + String(format: "%02x", $1) }
     }
