+++ KinoLenta/Common/Extensions/StringExtension.swift
@@ -1,33 +1,13 @@
 //
-//  StringExtension.swift
+//  Extension+String.swift
 //  KinoLenta
 //
-//  Created by Dmitry Trifonov on 10.12.2021.
+//  Created by user on 16.12.2021.
 //
 
 import Foundation
-import UIKit
+
 
 extension String {
-    func width(withHeight height: CGFloat, font: UIFont) -> CGFloat {
-        let constraintRect = CGSize(width: .greatestFiniteMagnitude, height: height)
-        let boundingBox = self.boundingRect(
-            with: constraintRect,
-            options: .usesLineFragmentOrigin,
-            attributes: [NSAttributedString.Key.font: font],
-            context: nil
-        )
-        return ceil(boundingBox.width)
-    }
-
-    func height(withWidth width: CGFloat, font: UIFont) -> CGFloat {
-        let rect = CGSize(width: width, height: .greatestFiniteMagnitude)
-        let boundingBox = self.boundingRect(
-            with: rect,
-            options: .usesLineFragmentOrigin,
-            attributes: [NSAttributedString.Key.font: font],
-            context: nil
-        )
-        return ceil(boundingBox.height)
-    }
+    var firstUppercased: String { prefix(1).uppercased() + dropFirst() }
 }
