+++ KinoLenta/Common/UI/CancellationHandle.swift
