+++ KinoLenta/Common/UI/CollectionViewCellDescriptor.swift
