+++ KinoLenta/Common/UI/Extensions/CollectionViewExtension.swift
