+++ KinoLenta/Common/UI/Extensions/DataUIExtension.swift
@@ -1,0 +1,8 @@
+import Foundation
+import UIKit
+
+extension Data {
+    func toImage() -> UIImage? {
+        UIImage(data: self)
+    }
+}
