+++ KinoLenta/Common/UI/Extensions/ImageIcons.swift
