+++ KinoLenta/Common/UI/ImageCollectionViewCell.swift
