+++ KinoLenta/Common/UI/ImageSizeTraits.swift
