+++ KinoLenta/Common/UI/LayoutManager.swift
