+++ KinoLenta/Common/UI/QuickItemFilter/QuickItemFilterCollectionViewCell.xib
