+++ KinoLenta/Common/UI/RatingView.swift
