+++ KinoLenta/Common/UI/StarsRatingDialogView.swift
